Array of mock points
