+++ src/main.js
@@ -6,12 +6,14 @@ import {createTripSort} from './view/trip-sort.js';
 import {createTripEventsList} from './view/events-list.js';
 import {createEditForm} from './view/edit-form.js';
 import {createTripEventsPoint} from './view/events-point.js';
-
-//import {generatePoint} from './mock/mock-point.js';
-import './mock/mock-point.js';
+import {generatePoint} from './mock/mock-point.js';
 
 
-const TRIP_POINTS = 3;
+const TRIP_POINTS = 15;
+
+const createMockPoints = (TRIP_POINTS) => new Array(TRIP_POINTS).fill(null).map(() => generatePoint());
+console.log(createMockPoints(TRIP_POINTS));
+
 
 const tripMainHeader = document.querySelector('.trip-main');
 const tripMenu = tripMainHeader.querySelector('.trip-controls');
@@ -33,7 +35,4 @@ for (let i=0; i < TRIP_POINTS; i++) {
 }
 
 // делаем два отдельных компонента: и для создания формы и для редактирования формы
-
-//generatePoint();
-
